requirements.txt
jupyter
cupy_cuda12x
h5py
matplotlib
numpy
scipy
tqdm
